Refactor institution classification fields and address layout
Moved `InstitutionClassificationFields` to a dedicated file for better modularity and reuse. Adjusted column widths in `InstitutionAddressInputForm` to improve layout and responsiveness. Updated imports in `DataDetails` accordingly.
